Navigation Fixes and layout adjustments
-Completely reworked the grid-template to be out of a 12 column system, and redid the rows a tad for mobile layouts as well.
-Fixed navigation on desktop viewport
-Removed menu-wrap class to reduce redundancy
- Prepared mobile layouts with beginning workings of hamburger menu.

Overall the navigation has been fixed, with a few things for the mobile viewport needing some final touches. I had realized as well that I was using a really restricted format for my grid layout, so I expanded it for versatility. Numerous fundamental style changes still need to be made like color choice, font styles, and adding the rest of the base content for things like the hero image.

The intended style is to match the whole schema with that of the resume page, so it keeps a cohesive look between the two, as it will be fully integrated.